refactor: moved status tab panel to own component
