+++ apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/ActiveJourneyList.spec.tsx
@@ -3,33 +3,30 @@ import { fireEvent, render, waitFor } from '@testing-library/react'
 import noop from 'lodash/noop'
 import { SnackbarProvider } from 'notistack'
 import { AuthUser } from 'next-firebase-auth'
-import { defaultJourney, oldJourney } from '../../journeyListData'
-import { ThemeProvider } from '../../../ThemeProvider'
-import { SortOrder } from '../../JourneySort'
+import { defaultJourney, oldJourney } from '../journeyListData'
+import { ThemeProvider } from '../../ThemeProvider'
+import { SortOrder } from '../JourneySort'
 import {
-  TrashedStatusTab,
-  GET_TRASHED_JOURNEYS,
-  RESTORE_TRASHED_JOURNEYS,
-  DELETE_TRASHED_JOURNEYS
-} from './TrashedStatusTab'
+  ActiveJourneyList,
+  ARCHIVE_ACTIVE_JOURNEYS,
+  GET_ACTIVE_JOURNEYS,
+  TRASH_ACTIVE_JOURNEYS
+} from './ActiveJourneyList'
 
-const trashedJourneysMock = {
+const activeJourneysMock = {
   request: {
-    query: GET_TRASHED_JOURNEYS
+    query: GET_ACTIVE_JOURNEYS
   },
   result: {
     data: {
-      journeys: [
-        { ...defaultJourney, trashedAt: '2021-12-07T03:22:41.135Z' },
-        { ...oldJourney, trashedAt: '2021-12-07T03:22:41.135Z' }
-      ]
+      journeys: [defaultJourney, oldJourney]
     }
   }
 }
 
 const noJourneysMock = {
   request: {
-    query: GET_TRASHED_JOURNEYS
+    query: GET_ACTIVE_JOURNEYS
   },
   result: {
     data: {
@@ -40,18 +37,13 @@ const noJourneysMock = {
 
 const authUser = { id: 'user-id1' } as unknown as AuthUser
 
-describe('TrashedStatusTab', () => {
-  beforeAll(() => {
-    jest.useFakeTimers('modern')
-    jest.setSystemTime(new Date('2021-12-11'))
-  })
-
+describe('ActiveJourneyList', () => {
   it('should render journeys in descending createdAt date by default', async () => {
     const { getAllByLabelText } = render(
-      <MockedProvider mocks={[trashedJourneysMock]}>
+      <MockedProvider mocks={[activeJourneysMock]}>
         <ThemeProvider>
           <SnackbarProvider>
-            <TrashedStatusTab onLoad={noop} event="" />
+            <ActiveJourneyList onLoad={noop} event="" />
           </SnackbarProvider>
         </ThemeProvider>
       </MockedProvider>
@@ -68,25 +60,21 @@ describe('TrashedStatusTab', () => {
   })
 
   it('should order journeys in alphabetical order', async () => {
-    const trashedLowerCaseJourneyTitle = {
+    const lowerCaseJourneyTitle = {
       ...defaultJourney,
-      title: 'a lower case title',
-      trashedAt: '2021-12-07T03:22:41.135Z'
+      title: 'a lower case title'
     }
-    const trashedOldJourney = {
-      ...oldJourney,
-      trashedAt: '2021-12-07T03:22:41.135Z'
-    }
+
     const { getAllByLabelText } = render(
       <MockedProvider
         mocks={[
           {
             request: {
-              query: GET_TRASHED_JOURNEYS
+              query: GET_ACTIVE_JOURNEYS
             },
             result: {
               data: {
-                journeys: [trashedLowerCaseJourneyTitle, trashedOldJourney]
+                journeys: [lowerCaseJourneyTitle, oldJourney]
               }
             }
           }
@@ -94,7 +82,7 @@ describe('TrashedStatusTab', () => {
       >
         <ThemeProvider>
           <SnackbarProvider>
-            <TrashedStatusTab
+            <ActiveJourneyList
               onLoad={noop}
               sortOrder={SortOrder.TITLE}
               event=""
@@ -114,43 +102,23 @@ describe('TrashedStatusTab', () => {
     )
   })
 
-  it('should exclude journeys older than 40 days', async () => {
-    const { getAllByLabelText } = render(
-      <MockedProvider
-        mocks={[
-          {
-            request: {
-              query: GET_TRASHED_JOURNEYS
-            },
-            result: {
-              data: {
-                journeys: [
-                  { ...defaultJourney, trashedAt: '2021-12-07T03:22:41.135Z' },
-                  { ...oldJourney, trashedAt: '2021-10-31T03:22:41.135Z' }
-                ]
-              }
-            }
-          }
-        ]}
-      >
+  it('should ask users to add a new journey', async () => {
+    const { getByRole, getByText } = render(
+      <MockedProvider mocks={[noJourneysMock]}>
         <ThemeProvider>
           <SnackbarProvider>
-            <TrashedStatusTab
-              onLoad={noop}
-              sortOrder={SortOrder.TITLE}
-              event=""
-            />
+            <ActiveJourneyList onLoad={noop} event="" />
           </SnackbarProvider>
         </ThemeProvider>
       </MockedProvider>
     )
-
     await waitFor(() =>
-      expect(getAllByLabelText('journey-card')[0].textContent).toContain(
-        'Default Journey Heading'
-      )
+      expect(getByText('No journeys to display.')).toBeInTheDocument()
     )
-    expect(getAllByLabelText('journey-card')[1]).toBeUndefined()
+    expect(
+      getByText('Create a journey, then find it here.')
+    ).toBeInTheDocument()
+    expect(getByRole('button')).toBeInTheDocument()
   })
 
   it('should render loading skeleton', async () => {
@@ -158,7 +126,7 @@ describe('TrashedStatusTab', () => {
       <MockedProvider mocks={[]}>
         <ThemeProvider>
           <SnackbarProvider>
-            <TrashedStatusTab onLoad={noop} event="" />
+            <ActiveJourneyList onLoad={noop} event="" />
           </SnackbarProvider>
         </ThemeProvider>
       </MockedProvider>
@@ -174,7 +142,7 @@ describe('TrashedStatusTab', () => {
       <MockedProvider mocks={[noJourneysMock]}>
         <ThemeProvider>
           <SnackbarProvider>
-            <TrashedStatusTab onLoad={onLoad} event="" />
+            <ActiveJourneyList onLoad={onLoad} event="" />
           </SnackbarProvider>
         </ThemeProvider>
       </MockedProvider>
@@ -182,13 +150,27 @@ describe('TrashedStatusTab', () => {
     await waitFor(() => expect(onLoad).toHaveBeenCalled())
   })
 
-  describe('Restore All', () => {
+  describe('Archive All', () => {
+    it('should display the archive all dialog', () => {
+      const { getByText } = render(
+        <MockedProvider mocks={[activeJourneysMock]}>
+          <ThemeProvider>
+            <SnackbarProvider>
+              <ActiveJourneyList onLoad={noop} event="archiveAllActive" />
+            </SnackbarProvider>
+          </ThemeProvider>
+        </MockedProvider>
+      )
+
+      expect(getByText('Archive Journeys')).toBeInTheDocument()
+    })
+
     const result = jest.fn(() => ({
-      data: [{ id: defaultJourney.id, status: 'published' }]
+      data: [{ id: defaultJourney.id, status: 'archived' }]
     }))
-    const restoreJourneysMock = {
+    const archiveJourneysMock = {
       request: {
-        query: RESTORE_TRASHED_JOURNEYS,
+        query: ARCHIVE_ACTIVE_JOURNEYS,
         variables: {
           ids: [defaultJourney.id, oldJourney.id]
         }
@@ -197,30 +179,16 @@ describe('TrashedStatusTab', () => {
     }
     const onLoad = jest.fn()
 
-    it('should display the restore all dialog', () => {
-      const { getByText } = render(
-        <MockedProvider mocks={[trashedJourneysMock]}>
-          <ThemeProvider>
-            <SnackbarProvider>
-              <TrashedStatusTab onLoad={noop} event="restoreAllTrashed" />
-            </SnackbarProvider>
-          </ThemeProvider>
-        </MockedProvider>
-      )
-
-      expect(getByText('Restore Journeys')).toBeInTheDocument()
-    })
-
-    it('should restore all journeys', async () => {
+    it('should archive all journeys', async () => {
       const { getByText } = render(
         <MockedProvider
-          mocks={[trashedJourneysMock, restoreJourneysMock, noJourneysMock]}
+          mocks={[activeJourneysMock, archiveJourneysMock, noJourneysMock]}
         >
           <ThemeProvider>
             <SnackbarProvider>
-              <TrashedStatusTab
+              <ActiveJourneyList
                 onLoad={onLoad}
-                event="restoreAllTrashed"
+                event="archiveAllActive"
                 authUser={authUser}
               />
             </SnackbarProvider>
@@ -228,25 +196,24 @@ describe('TrashedStatusTab', () => {
         </MockedProvider>
       )
       await waitFor(() => expect(onLoad).toHaveBeenCalled())
-      fireEvent.click(getByText('Restore'))
+      fireEvent.click(getByText('Archive'))
       await waitFor(() => expect(result).toHaveBeenCalled())
     })
 
-    // test intermittently fails due to snackbar and dom timeout
-    xit('should show error', async () => {
+    it('should show error', async () => {
       const { getByText } = render(
         <MockedProvider
           mocks={[
-            trashedJourneysMock,
-            { ...restoreJourneysMock, error: new Error('error') }
+            activeJourneysMock,
+            { ...archiveJourneysMock, error: new Error('error') }
           ]}
         >
           <SnackbarProvider>
             <ThemeProvider>
               <SnackbarProvider>
-                <TrashedStatusTab
+                <ActiveJourneyList
                   onLoad={onLoad}
-                  event="restoreAllTrashed"
+                  event="archiveAllActive"
                   authUser={authUser}
                 />
               </SnackbarProvider>
@@ -255,18 +222,18 @@ describe('TrashedStatusTab', () => {
         </MockedProvider>
       )
       await waitFor(() => expect(onLoad).toHaveBeenCalled())
-      fireEvent.click(getByText('Restore'))
+      fireEvent.click(getByText('Archive'))
       await waitFor(() => expect(getByText('error')).toBeInTheDocument())
     })
   })
 
-  describe('Delete All', () => {
+  describe('Trash All', () => {
     const result = jest.fn(() => ({
-      data: [{ id: defaultJourney.id, status: 'deleted' }]
+      data: [{ id: defaultJourney.id, status: 'archived' }]
     }))
-    const deleteJourneysMock = {
+    const trashJourneysMock = {
       request: {
-        query: DELETE_TRASHED_JOURNEYS,
+        query: TRASH_ACTIVE_JOURNEYS,
         variables: {
           ids: [defaultJourney.id, oldJourney.id]
         }
@@ -275,30 +242,30 @@ describe('TrashedStatusTab', () => {
     }
     const onLoad = jest.fn()
 
-    it('should display the delete all dialog', () => {
+    it('should display the trash all dialog', () => {
       const { getByText } = render(
-        <MockedProvider mocks={[trashedJourneysMock]}>
+        <MockedProvider mocks={[activeJourneysMock]}>
           <ThemeProvider>
             <SnackbarProvider>
-              <TrashedStatusTab onLoad={noop} event="deleteAllTrashed" />
+              <ActiveJourneyList onLoad={noop} event="trashAllActive" />
             </SnackbarProvider>
           </ThemeProvider>
         </MockedProvider>
       )
 
-      expect(getByText('Delete Journeys Forever')).toBeInTheDocument()
+      expect(getByText('Trash Journeys')).toBeInTheDocument()
     })
 
     it('should trash all journeys', async () => {
       const { getByText } = render(
         <MockedProvider
-          mocks={[trashedJourneysMock, deleteJourneysMock, noJourneysMock]}
+          mocks={[activeJourneysMock, trashJourneysMock, noJourneysMock]}
         >
           <ThemeProvider>
             <SnackbarProvider>
-              <TrashedStatusTab
+              <ActiveJourneyList
                 onLoad={onLoad}
-                event="deleteAllTrashed"
+                event="trashAllActive"
                 authUser={authUser}
               />
             </SnackbarProvider>
@@ -306,25 +273,24 @@ describe('TrashedStatusTab', () => {
         </MockedProvider>
       )
       await waitFor(() => expect(onLoad).toHaveBeenCalled())
-      fireEvent.click(getByText('Delete Forever'))
+      fireEvent.click(getByText('Trash'))
       await waitFor(() => expect(result).toHaveBeenCalled())
     })
 
-    // test intermittently fails due to snackbar and dom timeout
-    xit('should show error', async () => {
+    it('should show error', async () => {
       const { getByText } = render(
         <MockedProvider
           mocks={[
-            trashedJourneysMock,
-            { ...deleteJourneysMock, error: new Error('error') }
+            activeJourneysMock,
+            { ...trashJourneysMock, error: new Error('error') }
           ]}
         >
           <SnackbarProvider>
             <ThemeProvider>
               <SnackbarProvider>
-                <TrashedStatusTab
+                <ActiveJourneyList
                   onLoad={onLoad}
-                  event="deleteAllTrashed"
+                  event="trashAllActive"
                   authUser={authUser}
                 />
               </SnackbarProvider>
@@ -333,10 +299,8 @@ describe('TrashedStatusTab', () => {
         </MockedProvider>
       )
       await waitFor(() => expect(onLoad).toHaveBeenCalled())
-      fireEvent.click(getByText('Delete Forever'))
-      await waitFor(() => expect(getByText('error')).toBeInTheDocument(), {
-        timeout: 1500
-      })
+      fireEvent.click(getByText('Trash'))
+      await waitFor(() => expect(getByText('error')).toBeInTheDocument())
     })
   })
 })
