+++ apps/journeys-admin/src/components/JourneyList/ActiveJourneyList/index.ts
@@ -1,0 +1,1 @@
+export { ActiveJourneyList } from './ActiveJourneyList'
