+++ apps/journeys-admin/src/components/JourneyList/ArchivedJourneyList/index.ts
@@ -1,0 +1,1 @@
+export { ArchivedJourneyList } from './ArchivedJourneyList'
