--- apps/journeys-admin/src/components/JourneyList/StatusTabPanel/ActiveStatusTab/index.ts
@@ -1,1 +1,0 @@
-export { ActiveStatusTab } from './ActiveStatusTab'
