--- apps/journeys-admin/src/components/JourneyList/StatusTabPanel/ArchivedStatusTab/index.ts
@@ -1,1 +1,0 @@
-export { ArchivedStatusTab } from './ArchivedStatusTab'
