--- apps/journeys-admin/src/components/JourneyList/StatusTabPanel/TrashedStatusTab/index.ts
@@ -1,1 +1,0 @@
-export { TrashedStatusTab } from './TrashedStatusTab'
