+++ apps/journeys-admin/src/components/JourneyList/TrashedJourneyList/index.ts
@@ -1,0 +1,1 @@
+export { TrashedJourneyList } from './TrashedJourneyList'
